jhipster virsion update

+++ src/main/java/io/iansoft/blockchain/service/dto/FixerExchangeRateDTO.java
@@ -1,62 +1,49 @@
 package io.iansoft.blockchain.service.dto;
-
-
-import io.iansoft.blockchain.domain.enumeration.CurrencyType;
-
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.validation.constraints.NotNull;
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.time.ZonedDateTime;
-import java.util.Objects;
 
-/**
- * A DTO for the Kraken entity.
- */
 public class FixerExchangeRateDTO implements Serializable {
 
     private Long id;
 
     private String base;
 
-    private BigDecimal USD;
-    private BigDecimal JPY;
-    private BigDecimal BGN;
-    private BigDecimal CZK;
-    private BigDecimal DKK;
-    private BigDecimal GBP;
-    private BigDecimal HUF;
-    private BigDecimal PLN;
-    private BigDecimal RON;
-    private BigDecimal SEK;
-    private BigDecimal CHF;
-    private BigDecimal NOK;
-    private BigDecimal HRK;
-    private BigDecimal RUB;
-    private BigDecimal TRY;
-    private BigDecimal AUD;
-    private BigDecimal BRL;
-    private BigDecimal CAD;
-    private BigDecimal CNY;
-    private BigDecimal HKD;
-    private BigDecimal IDR;
-    private BigDecimal ILS;
-    private BigDecimal INR;
-    private BigDecimal KRW;
-    private BigDecimal MXN;
-    private BigDecimal MYR;
-    private BigDecimal NZD;
-    private BigDecimal PHP;
-    private BigDecimal SGD;
-    private BigDecimal THB;
-    private BigDecimal EUR;
+    private double USD;
+    private double JPY;
+    private double BGN;
+    private double CZK;
+    private double DKK;
+    private double GBP;
+    private double HUF;
+    private double PLN;
+    private double RON;
+    private double SEK;
+    private double CHF;
+    private double NOK;
+    private double HRK;
+    private double RUB;
+    private double TRY;
+    private double AUD;
+    private double BRL;
+    private double CAD;
+    private double CNY;
+    private double HKD;
+    private double IDR;
+    private double ILS;
+    private double INR;
+    private double KRW;
+    private double MXN;
+    private double MYR;
+    private double NZD;
+    private double PHP;
+    private double SGD;
+    private double THB;
+    private double EUR;
 
-    public BigDecimal getEUR() {
+    public double getEUR() {
         return EUR;
     }
 
-    public void setEUR(BigDecimal EUR) {
+    public void setEUR(double EUR) {
         this.EUR = EUR;
     }
 
@@ -76,243 +63,243 @@ public class FixerExchangeRateDTO implements Serializable {
         this.base = base;
     }
 
-    public BigDecimal getUSD() {
+    public double getUSD() {
         return USD;
     }
 
-    public void setUSD(BigDecimal USD) {
+    public void setUSD(double USD) {
         this.USD = USD;
     }
 
-    public BigDecimal getJPY() {
+    public double getJPY() {
         return JPY;
     }
 
-    public void setJPY(BigDecimal JPY) {
+    public void setJPY(double JPY) {
         this.JPY = JPY;
     }
 
-    public BigDecimal getBGN() {
+    public double getBGN() {
         return BGN;
     }
 
-    public void setBGN(BigDecimal BGN) {
+    public void setBGN(double BGN) {
         this.BGN = BGN;
     }
 
-    public BigDecimal getCZK() {
+    public double getCZK() {
         return CZK;
     }
 
-    public void setCZK(BigDecimal CZK) {
+    public void setCZK(double CZK) {
         this.CZK = CZK;
     }
 
-    public BigDecimal getDKK() {
+    public double getDKK() {
         return DKK;
     }
 
-    public void setDKK(BigDecimal DKK) {
+    public void setDKK(double DKK) {
         this.DKK = DKK;
     }
 
-    public BigDecimal getGBP() {
+    public double getGBP() {
         return GBP;
     }
 
-    public void setGBP(BigDecimal GBP) {
+    public void setGBP(double GBP) {
         this.GBP = GBP;
     }
 
-    public BigDecimal getHUF() {
+    public double getHUF() {
         return HUF;
     }
 
-    public void setHUF(BigDecimal HUF) {
+    public void setHUF(double HUF) {
         this.HUF = HUF;
     }
 
-    public BigDecimal getPLN() {
+    public double getPLN() {
         return PLN;
     }
 
-    public void setPLN(BigDecimal PLN) {
+    public void setPLN(double PLN) {
         this.PLN = PLN;
     }
 
-    public BigDecimal getRON() {
+    public double getRON() {
         return RON;
     }
 
-    public void setRON(BigDecimal RON) {
+    public void setRON(double RON) {
         this.RON = RON;
     }
 
-    public BigDecimal getSEK() {
+    public double getSEK() {
         return SEK;
     }
 
-    public void setSEK(BigDecimal SEK) {
+    public void setSEK(double SEK) {
         this.SEK = SEK;
     }
 
-    public BigDecimal getCHF() {
+    public double getCHF() {
         return CHF;
     }
 
-    public void setCHF(BigDecimal CHF) {
+    public void setCHF(double CHF) {
         this.CHF = CHF;
     }
 
-    public BigDecimal getNOK() {
+    public double getNOK() {
         return NOK;
     }
 
-    public void setNOK(BigDecimal NOK) {
+    public void setNOK(double NOK) {
         this.NOK = NOK;
     }
 
-    public BigDecimal getHRK() {
+    public double getHRK() {
         return HRK;
     }
 
-    public void setHRK(BigDecimal HRK) {
+    public void setHRK(double HRK) {
         this.HRK = HRK;
     }
 
-    public BigDecimal getRUB() {
+    public double getRUB() {
         return RUB;
     }
 
-    public void setRUB(BigDecimal RUB) {
+    public void setRUB(double RUB) {
         this.RUB = RUB;
     }
 
-    public BigDecimal getTRY() {
+    public double getTRY() {
         return TRY;
     }
 
-    public void setTRY(BigDecimal TRY) {
+    public void setTRY(double TRY) {
         this.TRY = TRY;
     }
 
-    public BigDecimal getAUD() {
+    public double getAUD() {
         return AUD;
     }
 
-    public void setAUD(BigDecimal AUD) {
+    public void setAUD(double AUD) {
         this.AUD = AUD;
     }
 
-    public BigDecimal getBRL() {
+    public double getBRL() {
         return BRL;
     }
 
-    public void setBRL(BigDecimal BRL) {
+    public void setBRL(double BRL) {
         this.BRL = BRL;
     }
 
-    public BigDecimal getCAD() {
+    public double getCAD() {
         return CAD;
     }
 
-    public void setCAD(BigDecimal CAD) {
+    public void setCAD(double CAD) {
         this.CAD = CAD;
     }
 
-    public BigDecimal getCNY() {
+    public double getCNY() {
         return CNY;
     }
 
-    public void setCNY(BigDecimal CNY) {
+    public void setCNY(double CNY) {
         this.CNY = CNY;
     }
 
-    public BigDecimal getHKD() {
+    public double getHKD() {
         return HKD;
     }
 
-    public void setHKD(BigDecimal HKD) {
+    public void setHKD(double HKD) {
         this.HKD = HKD;
     }
 
-    public BigDecimal getIDR() {
+    public double getIDR() {
         return IDR;
     }
 
-    public void setIDR(BigDecimal IDR) {
+    public void setIDR(double IDR) {
         this.IDR = IDR;
     }
 
-    public BigDecimal getILS() {
+    public double getILS() {
         return ILS;
     }
 
-    public void setILS(BigDecimal ILS) {
+    public void setILS(double ILS) {
         this.ILS = ILS;
     }
 
-    public BigDecimal getINR() {
+    public double getINR() {
         return INR;
     }
 
-    public void setINR(BigDecimal INR) {
+    public void setINR(double INR) {
         this.INR = INR;
     }
 
-    public BigDecimal getKRW() {
+    public double getKRW() {
         return KRW;
     }
 
-    public void setKRW(BigDecimal KRW) {
+    public void setKRW(double KRW) {
         this.KRW = KRW;
     }
 
-    public BigDecimal getMXN() {
+    public double getMXN() {
         return MXN;
     }
 
-    public void setMXN(BigDecimal MXN) {
+    public void setMXN(double MXN) {
         this.MXN = MXN;
     }
 
-    public BigDecimal getMYR() {
+    public double getMYR() {
         return MYR;
     }
 
-    public void setMYR(BigDecimal MYR) {
+    public void setMYR(double MYR) {
         this.MYR = MYR;
     }
 
-    public BigDecimal getNZD() {
+    public double getNZD() {
         return NZD;
     }
 
-    public void setNZD(BigDecimal NZD) {
+    public void setNZD(double NZD) {
         this.NZD = NZD;
     }
 
-    public BigDecimal getPHP() {
+    public double getPHP() {
         return PHP;
     }
 
-    public void setPHP(BigDecimal PHP) {
+    public void setPHP(double PHP) {
         this.PHP = PHP;
     }
 
-    public BigDecimal getSGD() {
+    public double getSGD() {
         return SGD;
     }
 
-    public void setSGD(BigDecimal SGD) {
+    public void setSGD(double SGD) {
         this.SGD = SGD;
     }
 
-    public BigDecimal getTHB() {
+    public double getTHB() {
         return THB;
     }
 
-    public void setTHB(BigDecimal THB) {
+    public void setTHB(double THB) {
         this.THB = THB;
     }
 }
